+++ Readme.md
@@ -1,7 +1,7 @@
 # Личный проект «Кексобукинг»
 
 * Студент: [Иван Ломов](https://up.htmlacademy.ru/javascript/22/user/1434441).
-* Наставник: `Ольга Поликашина`.
+* Наставник: [Ольга Поликашина](https://up.htmlacademy.ru/javascript/22/user/890283).
 
 ---
